nad_logging_service/requirements.txt
Flask
black
flake8
isort
pytest
pytest-mock
flask-cors
flask-limiter
python-dotenv
python-dateutil